+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 ## 1.11.1
 
 ### Bug Fixes
-- [KAFKA-390](https://jira.mongodb.org/browse/KAFKA-390) Logging incompatible properties no longer NPEs on null values
+  - [KAFKA-391](https://jira.mongodb.org/browse/KAFKA-391) Add wildcard support for starts with field name matching
+  - [KAFKA-390](https://jira.mongodb.org/browse/KAFKA-390) Logging incompatible properties no longer NPEs on null values
 
 ## 1.11.0